tidying up

+++ src/commands/bundle.js
@@ -6,20 +6,16 @@ module.exports = function (parameters, done) {
     var fs = require("fs-extra");
     var exec = require('../functions/exec');
 
+    if (!config.projectFound())
+        throw new Error("You're not inside a smallstack project folder!");
+
     fs.removeSync(path.join(config.builtDirectory, "meteor.tar.gz"));
 
     console.log("Creating bundle in directory : ", config.builtDirectory);
 
-    // if (!parameters.server)
-    //     throw new Error("Please set --server!");
-
-    // exec("meteor build " + path.relative(config.meteorDirectory, config.builtDirectory) + " --architecture os.linux.x86_64 --server " + parameters.server, {
-    //     cwd: config.meteorDirectory
-    // });
-
     exec("meteor build " + path.relative(config.meteorDirectory, config.builtDirectory) + " --architecture os.linux.x86_64 --server-only", {
         cwd: config.meteorDirectory,
-        finished: function() {
+        finished: function () {
             done();
         }
     });
